Migrate from tap port to all outputs from same port
With this change both the main stream and tap streams all connect
to the same output port from a node (application or channel).
The links will be solid or dotted where dotted indicates it is a
tap channel. There are new controls on a link to flip from a
tap channel to a main channel. There is also a new link control
to insert a channel between two elements (and wire them up
automatically).

Additionally the channel/destination representation has changed to
make it more distinct from a regular app.